Rename theming to styling for better readability
Renamed from theming to styling and grouped together the logic files in theme_manager to prepare for adding themes and color tokens
