+++ src/ert/gui/styling/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from ert.gui.styling.theme_manager.manager import ColorSchemeManager
+from ert.gui.styling.theme_manager.theme import ColorScheme
+
+__all__ = [
+    "ColorScheme",
+    "ColorSchemeManager",
+]
