+++ src/ert/gui/styling/theme_manager/manager.py
