+++ src/ert/gui/styling/theme_manager/theme.py
