--- src/ert/gui/theming/__init__.py
@@ -1,9 +1,0 @@
-from __future__ import annotations
-
-from .manager import ColorSchemeManager
-from .theme import ColorScheme
-
-__all__ = [
-    "ColorScheme",
-    "ColorSchemeManager",
-]
