+++ tests/ert/unit_tests/gui/styling/test_theme_detection.py
@@ -4,9 +4,9 @@ import pytest
 from PyQt6.QtCore import Qt
 from PyQt6.QtGui import QColor, QGuiApplication, QPalette
 
-from ert.gui.theming import ColorScheme
-from ert.gui.theming import manager as manager_module
-from ert.gui.theming.manager import (
+from ert.gui.styling import ColorScheme
+from ert.gui.styling.theme_manager import manager as manager_module
+from ert.gui.styling.theme_manager.manager import (
     _DARK_BASE_VALUE_THRESHOLD,
     detect_system_color_scheme,
 )
